mejora backend carrito arreglo de diseño carrito frontend

+++ server/src/controllers/carro.js
@@ -44,8 +44,29 @@ const eliminarDelCarrito = async (req, res) => {
     const { idCarrito } = req.params;
 
     try {
-        await db.query('DELETE FROM carrito WHERE id = ?', [idCarrito]);
-        res.status(200).json({ mensaje: 'Producto eliminado del carrito' });
+        // Primero obtenemos la cantidad del producto en el carrito
+        const [producto] = await db.query('SELECT cantidad FROM carrito WHERE id = ?', [idCarrito]);
+        
+        // Si no encontramos el producto
+        if (!producto) {
+            return res.status(404).json({ mensaje: 'Producto no encontrado en el carrito' });
+        }
+        console.log(producto);
+
+        // Si la cantidad es menor o igual a 0, eliminamos el producto
+        if (parseInt(producto[0].cantidad) <= 1) {
+            console.log("se cumple");
+            await db.query('DELETE FROM carrito WHERE id = ?', [idCarrito]);
+            return res.status(200).json({ mensaje: 'Producto eliminado del carrito' });
+        }else{
+            console.log("no se cumple");
+            await db.query('UPDATE carrito SET cantidad = cantidad - 1 WHERE id = ?', [idCarrito]);
+            return res.status(200).json({ mensaje: 'Cantidad del producto reducida en 1' });
+        }
+
+        // Si la cantidad es mayor a 0, restamos 1 de la cantidad
+     
+
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
